# U512 add/sub optimize