considering the triviality of implementation and impact, want to merge it as small separate pr

```
test u512_add ... bench:     172,925 ns/iter (+/- 6,794)
test u512_sub ... bench:   1,491,146 ns/iter (+/- 60,039)

test u512_add ... bench:      96,458 ns/iter (+/- 3,181)
test u512_sub ... bench:      70,440 ns/iter (+/- 3,269)
```